book_management/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.0
pytest-django==4.5.2
pytest==7.2.0
